Organize namespaces and folder structure
