+++ DotNetNakamaServer/MatchDispatcher.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json;
+using System.Text.Json;
+using DotNetNakamaServer.NetworkTransport;
 
 namespace DotNetNakamaServer;
