+++ DotNetNakamaServer/Matchmaking/MatchmakingService.cs
@@ -2,6 +2,8 @@ using System.Collections.Concurrent;
 using System.Text.Json;
 using DotNetNakamaServer;
 
+namespace DotNetNakamaServer.Matchmaking;
+
 public class MatchmakingService
 {
     private readonly MatchEngine _matchEngine;
