+++ DotNetNakamaServer/Matchmaking/MatchmakingTicket.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
@@ -6,6 +6,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Text.Json;
 using DotNetNakamaServer;
+
+namespace DotNetNakamaServer.Matchmaking;
 
 public class MatchmakingTicket
 {
