+++ DotNetNakamaServer/NetworkTransport/DisconnectReason.cs
@@ -1,4 +1,4 @@
-﻿namespace DotNetNakamaServer;
+namespace DotNetNakamaServer.NetworkTransport;
 
 public enum DisconnectReason
 {
